@@ -21,7 +21,7 @@
         "fideloper/proxy": "^4.0",
         "guzzlehttp/guzzle": "^6.3",
         "intervention/image": "^2.5",
-        "laravel/framework": "^6.0",
+        "laravel/framework": "6.5.2",
         "laravel/tinker": "^1.0",
         "laravelrus/sleepingowl": "dev-bs4",
         "mews/purifier": "^3.1",
